breaking change: remove frame.Sweep
The frame.Sweep method is impossible (or at least very difficult) to support in
a way that doesn't directly depend on exp/shiny. It also requires this package
to bring in exp/shiny as a dependency, which I don't want.

Sweep read mouse events through the shiny event pipe and expanded or
retracted the selection as those events were processed. It required the shiny
event pipe to function properly.

Here is the frameSweep method as a function, in it's entirety:

	func frameSweep(f *frame.Frame, ep screen.Window, flush func()) {
		p0, p1 := f.Dot()
		for {
			switch e := ep.NextEvent().(type) {
			case mouse.Event:
				if e.Direction != 0 {
					ep.SendFirst(e)
					return
				}
				if p1 = f.IndexOf(pt(e)); p0 > p1 {
					f.Select(p1, p0)
				} else {
					f.Select(p0, p1)
				}
				flush()
			case interface{}:
				ep.SendFirst(e)
				return
			}
		}
	}

Several things are wrong with this function. For one, it relies on the
event driven behavior that the original libframe did not support. (mouse
events were sent on a dedicated channel).

The last version of github.com/as/frame containing frame.Sweep was v0.6.12,
also tagged as "sweep". To maintain compatibility with programs or sources using
the sweep method, lock to v0.6.12 or below.

If you want to continue using the new version with exp/shiny, copy the function
above into your source and call it instead of f.Sweep. (It may be possible to
produce a go fix for this at a later time).