frame: point proximity for indexof
If the point is past 2/3 of the glyph, assume we want the next one. To
enable, set enablePointProximity to true.

Note: There are still some things wrong with this branch:

(1) indexof is buggy around spaces and tabs, especially when
mixed with regular text

(2) scrolling and rendering the entire frame doesn't work