frame: vendor dependencies
We vendor dependencies from the golang/freetype and golang/x
locations to increase build speeds. There are further dependencies
required to run the shiny examples in the examples/ directory. We
do not vendor those, as they are auxillary dependencies and outside
the scope of frame.

This adds approximately 3MiB to the repository, but the cost is worth
the benefit in my opinion for reproducible builds.